Wrapped Splashcreen around Material App in testing
