+++ test/features/splash/splash_screen_test.dart
@@ -3,36 +3,26 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:lottie/lottie.dart';
 import 'package:productive/common/resources/r.dart';
 import 'package:productive/features/splash/splash_screen.dart';
-import 'package:productive/main.dart';
 import 'package:productive/utils/animation/animated_scale.dart';
 
-import 'test_utils.dart';
+import '../../utils/widget/test_utils.dart';
 
 void main() {
-  group("Basic App Setup test", () {
-    testWidgets('App Follows MaterialTheme', (WidgetTester tester) async {
-      await tester.pumpWidget(ProductiveApp(
-        testChild: Container(),
-      ));
-      expect(find.byType(MaterialApp), findsOneWidget);
-    });
-  });
-
   group("Splash Screen Tests", () {
     testWidgets("There is Lottie Widget", (WidgetTester tester) async {
-      await tester.pumpWidget(const SplashScreenBody().testWidget);
+      await tester.pumpWidget(const SplashScreen().asScaffold);
       expect(find.byType(LottieBuilder), findsOneWidget);
     });
 
     testWidgets("There is a Productive Text Logo", (WidgetTester tester) async {
-      await tester.pumpWidget(const SplashScreenBody().testWidget);
+      await tester.pumpWidget(const SplashScreen().asScaffold);
       expect(find.byType(ShaderMask), findsOneWidget);
       expect(find.text(R.string.appName), findsOneWidget);
     });
 
     testWidgets("There is a Google Sign in Button",
         (WidgetTester tester) async {
-      await tester.pumpWidget(const SplashScreenBody().testWidget);
+      await tester.pumpWidget(const SplashScreen().asScaffold);
       expect(find.byType(AnimatedScale), findsOneWidget);
       expect(find.text(R.string.loginWithGoogle), findsOneWidget);
     });
