--- test/unit_test.dart
@@ -1,9 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  group("blablabla", () {
-    test("unit tests", () {
-      expect(5, 4 + 1);
-    });
-  });
-}
