+++ test/utils/widget/test_utils.dart
@@ -3,4 +3,6 @@ import 'package:flutter/material.dart';
 extension TestWidget on Widget {
   Widget get testWidget =>
       Directionality(textDirection: TextDirection.ltr, child: this);
+
+  Widget get asScaffold => MaterialApp(home: this);
 }
